requirements.txt
flask==1.1.2
tensorflow==2.3.0
Flask-RESTful==0.3.8
numpy==1.18.5
opencv-python==4.5.1.48
pillow~=8.1.0